+++ tests/com/yanncebron/m68kplugin/parser/DirectivesParsingTest.java
@@ -530,6 +530,10 @@ public class DirectivesParsingTest extends M68kParsingTestCase {
     doCodeTest(" machine 68000");
   }
 
+  public void testMachineDirectiveId() throws Exception {
+    doCodeTest(" machine mc68000");
+  }
+
   public void testMachineDirectiveMissingCpuType() throws Exception {
     doCodeTest(" machine ");
   }